connected firebase functions

+++ src/features/firestoreCart.js
@@ -1,46 +1,57 @@
-import { collection, doc, addDoc, getDocs, query, where, deleteDoc } from "firebase/firestore"
-import { db } from "../firebase"
-import { useSelector } from "react-redux"
+import {
+  collection,
+  doc,
+  addDoc,
+  getDocs,
+  setDoc,
+  query,
+  where,
+  deleteDoc,
+} from "firebase/firestore";
+import { db, auth } from "../firebase";
 
-export const uploadCart = (cartItems) => {
-    const currentUser = useSelector(state => state.user.currentUser);
+export const uploadItem = (movie) => {
+  const currentUser = auth.currentUser;
+  if (currentUser) {
     const uid = currentUser.uid;
+    const collectionRef = collection(db, "users", uid, "cart");
+
+    addDoc(collectionRef, movie);
+  }
+};
+
+export const removeItem = async (movie) => {
+  const currentUser = auth.currentUser;
+  if (currentUser) {
+    console.log("attempting delete")
+    const uid = currentUser.uid;
+    const movie_id = movie.id;
 
     const collectionRef = collection(db, "users", uid, "cart");
 
-    cartItems.forEach(cartItem => {
-        addDoc(collectionRef, cartItem);
-    });
-
-}
-
-export const removeItem = async (movie_id) => {
-    const currentUser = useSelector(state => state.user.currentUser);
-    const uid = currentUser.uid;
-
-    const collectionRef = collection(db, "users", uid, "cart");
-
-    const query = query(collectionRef, where("movie_id", "==", movie_id));
-
-    const querySnapshot = await getDocs(query);
+    const querySnapshot = await getDocs(query(collectionRef, where("movie_id", "==", movie_id)));
     querySnapshot.forEach((doc) => {
-        deleteDoc(doc.ref);
+      deleteDoc(doc.ref);
     });
-}
+  }
+};
 
 export const emptyCart = async () => {
-    const currentUser = useSelector(state => state.user.currentUser);
+  const currentUser = auth.currentUser;
+  if (currentUser) {
     const uid = currentUser.uid;
 
     const collectionRef = collection(db, "users", uid, "cart");
     const querySnapshot = await getDocs(collectionRef);
     querySnapshot.forEach((doc) => {
-        deleteDoc(doc.ref);
+      deleteDoc(doc.ref);
     });
-}
+  }
+};
 
 export const getCart = async () => {
-    const currentUser = useSelector(state => state.user.currentUser);
+  const currentUser = auth.currentUser;
+  if (currentUser) {
     const uid = currentUser.uid;
 
     const collectionRef = collection(db, "users", uid, "cart");
@@ -52,6 +63,7 @@ export const getCart = async () => {
       // doc.data() is never undefined for query doc snapshots
       cartItems.push(doc.data());
     });
-
+    
     return cartItems;
-}
+  }
+};
